use u8 for reference counter

+++ src/browser/webapi/net/FormData.zig
@@ -34,7 +34,7 @@ const Allocator = std.mem.Allocator;
 
 const FormData = @This();
 
-_rc: lp.RC(u32),
+_rc: lp.RC(u8),
 
 _arena: Allocator,
 _entries: std.ArrayList(Entry),
